Split Result Modal component
